Fix 'undefined index' warning in the pivot function of the array helper.
This bug is related to https://bugs.php.net/bug.php?id=66173 and to a strange
behaviour which also can be seen here https://3v4l.org/tofaS .
In short words: We shouldn't use numeric object properties at all, because they
start to act crazy if cast to an array. This commit fixes at least the pivot function,
a better fix would be to refactor the whole FAF workflow to not use numeric properties,
but this is too much work for the moment.